Add rewriter abstraction and JSON-API rewriting
Abstracted out the idea of a "rewriter", whose job it is to know how
to change incoming mutations based on previous operations. I've
currently only added a JSON-API rewriter that knows how to traverse
JSON-API data and update IDs.

I found there was quite a bit of messing around with the JSON body
values. Would be good at some point to go through and clean this up.